refactor(langy): extract 20 inline tools into per-entity modules (PR-4.0)
Phase 4 prep — no behavior change.

`langy.ts` had ~900 LOC of inline tool definitions (lines 250-1169
before this commit). Mastra's tool-runtime model expects per-tool
modules with a context object rather than a single closure capturing
projectId/seenIds/services from the enclosing function. Doing this
extraction now means PR-4.3 (the Mastra spike) can drop the same
tool factories into a Mastra Agent definition unchanged.

New shape:
- `src/server/services/langy/tools/types.ts` — `LangyToolContext`
  (projectId, experimentSlug, evaluatorService, promptService, seenIds,
  prisma). Per-request value threaded through every factory.
- `tools/evaluators.ts` — 6 tools (list/get + 4 propose_*)
- `tools/prompts.ts` — 5 tools (list/get/search + propose_create/update)
- `tools/datasets.ts` — 4 tools (list/get + propose_create/add_rows)
- `tools/workbench.ts` — 3 tools (get_state, find_failing_rows,
  propose_run) + the helpers (summarizeWorkbenchState,
  summarizePerTargetResults, findFailingRows, extractRowInputs)
- `tools/traces.ts` — search_traces
- `tools/runs.ts` — search_past_runs
- `tools/index.ts` — `buildLangyTools(ctx)` barrel assembles all 20
  into the map the route hands to `streamText`. Mastra wires through
  the same factory shape in PR-4.3.

`langy.ts` shrunk by ~1130 lines: imports cleaned up (dropped `tool`,
`z`, AVAILABLE_EVALUATORS, getEvaluator helpers, parseEvaluationResult,
persistedEvaluationsV3StateSchema, esClient, TRACE_INDEX — all moved
to the modules that actually use them), the inline tools block became
one `buildLangyTools({...})` call, and the four helper functions at
the bottom of the file are gone (now in tools/workbench.ts).

The tool-surface contract test (`langy.tool-surface.unit.test.ts`)
used to regex-scan `langy.ts` source for `tool({`. It now instantiates
`buildLangyTools` with a stub context and inspects the returned keys —
authoritative source instead of regex over moving source.

Test surface unchanged: 117/117 unit + 43/43 integration green (the
2 pre-existing ProposalCard label-assertion failures predate this
commit and predate Phase 3 — separate issue).
